requirements.txt
pillow
requests
wikipedia-api
pytest
flake8
mypy
playsound==1.3.0
